src/multistackcluster/nompirequirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.13=hb9d3cd8_0
aom=3.9.1=hac33072_0
astroid=3.3.5=py311h38be061_0
asttokens=3.0.0=pyhd8ed1ab_1
attr=2.5.1=h166bdaf_1
autograd=1.7.0=pyhd8ed1ab_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.2=heb4867d_0
ca-certificates=2025.1.31=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.18.2=h3394656_1
certifi=2025.1.31=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.0=py311hd18a35c_2
cycler=0.12.1=pyhd8ed1ab_0
cyrus-sasl=2.1.27=h54b06d7_7
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
debugpy=1.8.7=py311hfdbb021_0
decorator=5.1.1=pyhd8ed1ab_1
dill=0.3.9=pyhd8ed1ab_0
double-conversion=3.3.0=h59595ed_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h5888daf_0
ffmpeg=7.1.0=gpl_h099772d_709
fftw=3.3.10=nompi_hf1063bd_110
flake8=7.1.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.1=py311h9ecbd09_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.12=hb9ae30d_0
gmp=6.3.0=hac33072_2
graphite2=1.3.13=h59595ed_1003
gsl=2.7=he838d99_0
h5py=3.10.0=nompi_py311h3839ddf_100
h5utils=1.13.2=nompi_ha1fc32c_1102
harfbuzz=10.2.0=h4bba637_0
harminv=1.4.2=hc567f7b_0
hdf5=1.14.2=nompi_h4f84152_100
icu=75.1=he02047a_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.32.0=pyh907856f_0
isort=5.13.2=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py311hd18a35c_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_1
lerc=4.0.0=h27087fc_0
level-zero=1.20.2=h84d6215_0
libabseil=20240722.0=cxx17_h5888daf_1
libaec=1.1.3=h59595ed_0
libass=0.17.3=hba53ac1_1
libblas=3.9.0=24_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcap=2.71=h39aace5_0
libcblas=3.9.0=24_linux64_openblas
libclang-cpp19.1=19.1.7=default_hb5137d0_0
libclang13=19.1.7=default_h9c6a7e4_0
libctl=4.5.1=h7f98852_0
libcups=2.3.3=h4637d8d_4
libcurl=8.10.1=hbbe4b11_0
libdeflate=1.22=hb9d3cd8_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_1
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libexpat=2.6.4=h5888daf_0
libfabric=2.0.0=ha770c72_1
libfabric1=2.0.0=h14e6f36_1
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgcrypt-lib=1.11.0=hb9d3cd8_2
libgdsii=0.21=h4bd325d_4
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_1
libglib=2.82.2=h2ff4ddf_1
libglvnd=1.7.0=ha4b6fd6_1
libglx=1.7.0=ha4b6fd6_1
libgomp=14.2.0=h77fa898_1
libgpg-error=1.51=hbd13f7d_1
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=24_linux64_openblas
libllvm19=19.1.7=ha7bfdaf_0
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnghttp2=1.58.0=h47da74e_1
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libntlm=1.4=h7f98852_1002
libopenblas=0.3.27=pthreads_hac2b453_1
libopengl=1.7.0=ha4b6fd6_1
libopenvino=2024.6.0=hac27bb2_3
libopenvino-auto-batch-plugin=2024.6.0=h4d9b6c2_3
libopenvino-auto-plugin=2024.6.0=h4d9b6c2_3
libopenvino-hetero-plugin=2024.6.0=h3f63f65_3
libopenvino-intel-cpu-plugin=2024.6.0=hac27bb2_3
libopenvino-intel-gpu-plugin=2024.6.0=hac27bb2_3
libopenvino-intel-npu-plugin=2024.6.0=hac27bb2_3
libopenvino-ir-frontend=2024.6.0=h3f63f65_3
libopenvino-onnx-frontend=2024.6.0=h6363af5_3
libopenvino-paddle-frontend=2024.6.0=h6363af5_3
libopenvino-pytorch-frontend=2024.6.0=h5888daf_3
libopenvino-tensorflow-frontend=2024.6.0=h630ec5c_3
libopenvino-tensorflow-lite-frontend=2024.6.0=h5888daf_3
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpmix=5.0.6=h658e747_0
libpng=1.6.44=hadc24fc_0
libpq=17.2=h3b95a9b_1
libprotobuf=5.28.3=h6128344_1
librsvg=2.58.4=h49af25d_2
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.48.0=hee588c1_1
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libsystemd0=257.2=h3dc2cb9_0
libtiff=4.7.0=he137b08_1
libudev1=257.3=h9a4d06a_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.14.1=hac33072_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=h8d12d68_1
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
lz4-c=1.10.0=h5888daf_1
markupsafe=3.0.2=py311h2dc5d0c_0
matplotlib=3.9.2=py311h38be061_1
matplotlib-base=3.9.2=py311h2b939e6_1
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mccabe=0.7.0=pyhd8ed1ab_0
mpb=1.11.1=nompi_h2e71656_104
mpi=1.0=openmpi
mpi4py=4.0.3=py311ha982e2a_0
mpld3=0.5.10=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_1
mysql-libs=9.0.1=he0572af_1
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pyhd8ed1ab_1
nlopt=2.8.0=py311hef2dd6d_2
numpy=1.26.4=py311h64a7726_0
ocl-icd=2.3.2=hd590300_1
openh264=2.5.0=hf92e6e3_0
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.9=he970967_0
openmpi=5.0.7=hb85ec53_100
openssl=3.4.1=h7b32b05_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=py311h7db5c69_1
pango=1.54.0=h861ebed_4
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.44=hba22ea6_2
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py311h1322bbf_0
pip=24.2=pyh8b19718_1
pixman=0.44.2=h29eaf8c_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.50=pyha770c72_0
psutil=6.1.1=py311h9ecbd09_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pugixml=1.14=h59595ed_0
pure_eval=0.2.3=pyhd8ed1ab_1
pycodestyle=2.12.1=pyhd8ed1ab_0
pyflakes=3.2.0=pyhd8ed1ab_0
pygments=2.19.1=pyhd8ed1ab_0
pylint=3.3.1=pyhd8ed1ab_0
pymeep=1.29.0=nompi_py311h1234567_103
pymeep-extras=1.29.0=nompi_py311h82cca05_103
pyparsing=3.2.0=pyhd8ed1ab_1
pyside6=6.8.1=py311h9053184_0
python=3.11.10=hc5c86c4_3_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.2=pyhd8ed1ab_0
python_abi=3.11=5_cp311
pytz=2024.1=pyhd8ed1ab_0
pyzmq=26.2.1=py311h7deb3e3_0
qhull=2020.2=h434a139_5
qt6-main=6.8.1=h588cce1_2
rdma-core=56.0=h5888daf_0
readline=8.2=h8228510_1
scipy=1.14.1=py311he1f765f_0
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.2.1=ha2e4443_0
stack_data=0.6.3=pyhd8ed1ab_1
svt-av1=2.3.0=h5888daf_0
tbb=2022.0.0=hceb3a55_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.2=pyhd8ed1ab_0
tomlkit=0.13.2=pyha770c72_0
tornado=6.4.1=py311h9ecbd09_1
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
ucc=1.3.0=h2b97398_4
ucx=1.18.0=hfd9a62f_1
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.44.0=pyhd8ed1ab_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.5=he73a12e_0
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.11=hb9d3cd8_1
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.6=ha6fb4c9_0
